new flag: infoOnly, will not generate screenshot, only get meta data of the article. for collect article info

+++ src/scan_article.py
@@ -120,27 +120,31 @@ def processPage(driver, config):
 
     # cleanup the article page
     config['meta'] = meta
-    nomalizeArticle(driver, config)
+    if not 'infoOnly' in config['article_info']:
+        nomalizeArticle(driver, config)
+    else:
+        logger.warning('meta will be saved')
 
     if 'debug' in config:
         input('finished clenup...')
 
     # get screenshot image
-    img = getScreenshot(driver, config)
-    if img == None:
-        logger.warning('failed to generate article image.')
-        return
+    if not 'infoOnly' in config['article_info']:
+        img = getScreenshot(driver, config)
+        if img == None:
+            logger.warning('failed to generate article image.')
+            return
 
-    # save page to outbox
-    fn = datetime.now().strftime('%Y%m%d-%H%M%S_' + config['name'] + '.jpg')
-    if 'contacts' in config:
-        copyToContacts(img, fn, config['contacts'])
-    else:
-        toFile = './outbox/' + fn
-        logger.info('generated file: %s', toFile)
-        import shutil
-        shutil.copyfile(img, toFile)
-    os.remove(img)
+        # save page to outbox
+        fn = datetime.now().strftime('%Y%m%d-%H%M%S_' + config['name'] + '.jpg')
+        if 'contacts' in config:
+            copyToContacts(img, fn, config['contacts'])
+        else:
+            toFile = './outbox/' + fn
+            logger.info('generated file: %s', toFile)
+            import shutil
+            shutil.copyfile(img, toFile)
+        os.remove(img)
     history.save(meta)
 
 def main(config):
